_to_be_deleted/[PJ] AI Co-Pilot ジギョケイ策定支援システム/04_Archive/Snapshots/Snapshot_2025-10-04T03-03-54-152Z/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
google-cloud-pubsub